requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama==0.4.4
cymem==2.0.6
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.3.0/en_core_web_trf-3.3.0-py3-none-any.whl
filelock==3.7.0
Flask==2.1.2
huggingface-hub==0.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.9
pyspellchecker==0.6.3
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-transformers==1.1.5
srsly==2.4.3
thinc==8.0.15
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.17.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.2
